Condenses .c files into 4 files, also condensed header files into 1 file
