--- src/execution/redirection.c
@@ -1,38 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-#include <fcntl.h>
-#include <string.h>
-
-void handle_redirection(char **args) {
-    for (int i = 0; args[i] != NULL; i++) {
-        
-        // Output Redirection - Overwrite ">"
-        if (strcmp(args[i], ">") == 0) {
-            // printf("DEBUG: Found output redirection to %s\n", args[i+1]);
-            int fd = open(args[i + 1], O_WRONLY | O_CREAT | O_TRUNC, 0644);
-            if (fd < 0) { perror("open"); exit(1); }
-            dup2(fd, STDOUT_FILENO);
-            close(fd);
-            args[i] = NULL; // "cut" the array here so execvp only sees the command
-        }
-
-        // Output Redirection - Append ">>"
-        else if (strcmp(args[i], ">>") == 0) {
-            int fd = open(args[i + 1], O_WRONLY | O_CREAT | O_APPEND, 0644);
-            if (fd < 0) { perror("open"); exit(1); }
-            dup2(fd, STDOUT_FILENO);
-            close(fd);
-            args[i] = NULL;
-        }
-
-        // Input Redirection "<"
-        else if (strcmp(args[i], "<") == 0) {
-            int fd = open(args[i + 1], O_RDONLY);
-            if (fd < 0) { perror("open"); exit(1); }
-            dup2(fd, STDIN_FILENO);
-            close(fd);
-            args[i] = NULL;
-        }
-    }
-}
